packages/traj/pyproject.toml
[project]
name = "aau-traj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kasper Fromm Pedersen", email = "kasperf@cs.aau.dk" }
]
requires-python = ">=3.13"
dependencies = [
    "adbc-driver-gizmosql==1.1.*",
    "duckdb==1.5.*",
    "jinja2==3.*",
    "loguru==0.7.*",
    "pydantic-settings==2.*",
    "typer==0.24.*"
]

[dependency-groups]
dev = [
  "pyarrow-stubs==20.0.*"
]


[project.scripts]
lakehouse = "aau_traj.main:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"
